тестовый буфер отправляется идеально.что-то не того с dma ацп

+++ user/include/main.h
@@ -28,6 +28,6 @@
 volatile uint32_t ui32ticksDelay;
 volatile uint32_t ticks_delay;
 
-void sendHello(uint32_t *pbuf, uint32_t buf_len);
+void sendHello(uint8_t *pbuf, uint32_t buf_len);
 
 #endif /* USER_INCLUDE_MAIN_H_ */
